test: PH2-T08 relocate HTTP and facade tests
Context:

- Phase: PH2
- Task: PH2-T08 Relocate and Adjust Tests Across Crates

Changes:

- Moved HTTP unit tests and fixture helpers from manja/src/kite/connect/**
  into manja-http/src/api/** and manja-http/src/test_utils.rs
- Added mockito as a dev-dependency in manja-http and gated its test_utils
  module behind #[cfg(test)] for crate-local test usage
- Added ticker unit tests in manja-ticker for TickerRequest JSON shape and
  StreamState/SubscriptionStream URI and message generation
- Introduced facade-level integration tests in manja/tests/facade_http.rs and
  manja/tests/facade_ticker.rs plus shared helpers in manja/tests/support.rs
- Removed legacy HTTP test modules and test_utils from the manja crate so HTTP
  behavior is unit-tested only in manja-http and facade tests use ManjaClient
- Verified workspace test coherence so cargo test --all runs HTTP, ticker, and
  facade tests without duplication

Notes:

- Tests:
  - cargo test -p manja-http
  - cargo test -p manja-ticker
  - cargo test -p manja
  - cargo test --all
- User/profile and margins fixtures remain missing; corresponding user API
  tests are kept commented in manja-http::api::user for future re-enabling
