+++ manja/src/kite/connect/api/portfolio.rs
@@ -143,122 +143,3 @@ impl<'c> Portfolio<'c> {
     // }
     // }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-
-    use mockito::ServerGuard;
-    use tokio::join;
-
-    use crate::kite::connect::client::test_utils::{
-        add_mocks, get_manja_test_client, read_to_object, APIEndpoint, HTTPMethod, TestResponse,
-    };
-    use crate::kite::connect::client::HTTPClient;
-    use crate::kite::connect::config::{Config, KITECONNECT_API_LOGIN, KITECONNECT_API_REDIRECT};
-    use crate::kite::connect::credentials::KiteCredentials;
-    use crate::kite::connect::models::{Auction, Holding};
-    use crate::kite::error::{KiteApiException, ManjaError};
-    use crate::test_support::init_tracing;
-
-    use super::*;
-
-    fn mock_map() -> HashMap<(HTTPMethod, APIEndpoint), TestResponse> {
-        let mut mmap = HashMap::new();
-        mmap.insert(
-            ("GET", "/portfolio/holdings"),
-            "./kiteconnect-mocks/holdings.json",
-        );
-        mmap.insert(
-            ("GET", "/portfolio/holdings/auctions"),
-            "./kiteconnect-mocks/auctions_list.json",
-        );
-        mmap
-    }
-
-    #[tokio::test]
-    async fn test_get_holdings_success() {
-        init_tracing();
-        let (server, mut manja_client) = get_manja_test_client().await;
-        let server_ptr: *const ServerGuard = &server;
-        log::debug!("Server @address: {:p}", server_ptr);
-        let (_server,) = join!(add_mocks(server, mock_map()));
-
-        let portfolio_api = Portfolio::new(&manja_client);
-        let response = portfolio_api.get_holdings().await.unwrap();
-        let expected =
-            read_to_object::<Vec<Holding>>("./kiteconnect-mocks/holdings.json").unwrap();
-
-        let holdings = response.data.expect("expected holdings data");
-        assert_eq!(holdings.len(), expected.len());
-        assert_eq!(holdings[0].tradingsymbol, expected[0].tradingsymbol);
-    }
-
-    #[tokio::test]
-    async fn test_get_auctions_success() {
-        let (server, mut manja_client) = get_manja_test_client().await;
-        let (_server,) = join!(add_mocks(server, mock_map()));
-
-        let portfolio_api = Portfolio::new(&manja_client);
-        let response = portfolio_api.get_auctions().await.unwrap();
-        let expected =
-            read_to_object::<Vec<Auction>>("./kiteconnect-mocks/auctions_list.json").unwrap();
-
-        let auctions = response.data.expect("expected auctions data");
-        assert_eq!(auctions.len(), expected.len());
-        assert_eq!(auctions[0].tradingsymbol, expected[0].tradingsymbol);
-    }
-
-    #[tokio::test]
-    async fn test_get_holdings_token_exception_error() {
-        init_tracing();
-        let mut server = mockito::Server::new_async().await;
-        let credentials = KiteCredentials::new(
-            "TEST_API_KEY",
-            "TEST_API_SECRET",
-            "TEST_USER_ID",
-            "TEST_PASSWORD",
-            "TEST_TOTP",
-        );
-        let config = Config::from_parts(
-            server.url(),
-            KITECONNECT_API_LOGIN.to_string(),
-            KITECONNECT_API_REDIRECT.to_string(),
-            credentials,
-        );
-        let mut client = HTTPClient::with_config(config);
-
-        let _m = server
-            .mock("GET", "/portfolio/holdings")
-            .with_status(403)
-            .with_header("content-type", "application/json")
-            .with_body(
-                r#"{
-                    "status": "error",
-                    "data": null,
-                    "message": "Token is invalid or has expired",
-                    "error_type": "TokenException"
-                }"#,
-            )
-            .create_async()
-            .await;
-
-        let portfolio_api = Portfolio::new(&client);
-        let err = portfolio_api
-            .get_holdings()
-            .await
-            .expect_err("expected token exception error");
-
-        match err {
-            ManjaError::KiteApiError(api_err) => {
-                assert_eq!(api_err.status_code, 403);
-                assert!(matches!(
-                    api_err.error_type,
-                    KiteApiException::TokenException
-                ));
-                assert_eq!(api_err.error_type.as_str(), "TokenException");
-            }
-            other => panic!("unexpected error variant: {:?}", other),
-        }
-    }
-}
