+++ manja/src/kite/connect/client.rs
@@ -464,97 +464,3 @@ fn http_config_from_manja_config(config: &Config) -> manja_http::Config {
         http_creds,
     )
 }
-
-#[cfg(test)]
-pub mod test_utils {
-    use std::collections::HashMap;
-
-    use super::*;
-
-    use mockito::{Mock, ServerGuard};
-    use crate::kite::connect::config::{Config, KITECONNECT_API_LOGIN, KITECONNECT_API_REDIRECT};
-    use crate::kite::connect::credentials::KiteCredentials;
-
-    pub fn resolve_fixture_path(relative: &str) -> std::path::PathBuf {
-        let root = std::env::var("CARGO_MANIFEST_DIR").unwrap_or_else(|_| ".".to_string());
-        let candidate = std::path::Path::new(&root).join(relative);
-        if candidate.exists() {
-            return candidate;
-        }
-
-        // When running in a workspace, fixture files may live at the workspace
-        // root (one level up from the crate).
-        let workspace_candidate = std::path::Path::new(&root).join("..").join(relative);
-        if workspace_candidate.exists() {
-            return workspace_candidate;
-        }
-
-        candidate
-    }
-
-    pub fn read_to_object<M>(path: &str) -> Result<M>
-    where
-        M: DeserializeOwned,
-    {
-        let full_path = resolve_fixture_path(path);
-        let contents = std::fs::read_to_string(&full_path).unwrap_or_else(|err| {
-            panic!("failed to read fixture at {}: {}", full_path.display(), err)
-        });
-        let obj: KiteApiResponse<M> = serde_json::from_str(&contents)?;
-        obj.data
-            .ok_or(ManjaError::Internal(format!("obj not found")))
-    }
-
-    pub type HTTPMethod = &'static str;
-    pub type APIEndpoint = &'static str;
-    pub type TestResponse = &'static str;
-
-    pub async fn add_mocks(
-        mut server: ServerGuard,
-        mock_map: HashMap<(HTTPMethod, APIEndpoint), TestResponse>,
-    ) -> ServerGuard {
-        let mut mocks = Vec::new();
-        for ((method, api_endpoint), response_path) in mock_map {
-            let full_path = resolve_fixture_path(response_path);
-            let response_json = std::fs::read_to_string(&full_path).unwrap_or_else(|err| {
-                panic!("failed to read fixture at {}: {}", full_path.display(), err)
-            });
-            let m = server
-                .mock(method, api_endpoint)
-                .with_status(200)
-                .with_header("content-type", "application/json")
-                .with_body(response_json)
-                .create_async();
-            mocks.push(m)
-        }
-        let _ms = futures::future::join_all(mocks).await;
-        server
-    }
-
-    pub async fn get_manja_test_client() -> (ServerGuard, HTTPClient) {
-        let server = mockito::Server::new_async().await;
-        // Load env vars (if enabled via feature)
-        #[cfg(feature = "dotenv-config")]
-        {
-            dotenv::dotenv().ok();
-        }
-        // Build a dedicated Config that points HTTPClient at the mock server.
-        let credentials = KiteCredentials::new(
-            "TEST_API_KEY",
-            "TEST_API_SECRET",
-            "TEST_USER_ID",
-            "TEST_PASSWORD",
-            "TEST_TOTP",
-        );
-        let config = Config::from_parts(
-            server.url(),
-            KITECONNECT_API_LOGIN.to_string(),
-            KITECONNECT_API_REDIRECT.to_string(),
-            credentials,
-        );
-        let session =
-            read_to_object::<UserSession>("./kiteconnect-mocks/generate_session.json").unwrap();
-
-        (server, HTTPClient::with_config(config).with_user_session(session))
-    }
-}
